PouchDB instead of LocalForage
