+++ package.json
@@ -32,6 +32,8 @@
     "hammerjs": "^2.0.8",
     "jssip": "^3.0.8",
     "localforage": "^1.5.0",
+    "pouchdb": "^6.3.4",
+    "pouchdb-find": "^6.3.4",
     "rxjs": "^5.4.1",
     "zone.js": "^0.8.14"
   },
