+++ src/app/directory/directory.component.html
@@ -1,5 +1,5 @@
 
-<h3 *ngIf="user && user.id" class="my-number">
+<h3 *ngIf="user && user._id" class="my-number">
   My number <span>{{ user.user | phoneNumber }}</span>
 </h3>
